requirements.txt
transformers==4.38.1
torch
faiss-cpu
PyPDF2
